requirements.txt
anvil-uplink==0.4.2
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
confection==0.1.3
cymem==2.0.8
filelock==3.12.4
fsspec==2023.9.2
future==0.18.3
gdown==5.1.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
mysql-client==0.0.1
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.0.3
pathy==0.10.2
pipdeptree==2.14.0
preshed==3.0.9
protobuf==3.20.3
pydantic==1.10.13
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.5.0
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.3.2
srsly==2.4.8
sympy==1.12
thinc==8.1.12
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
triton==2.1.0
typer==0.7.0
typing-extensions==4.8.0
tzdata==2024.1
urllib3==2.0.6
wasabi==1.1.2
ws4py==0.5.1
